\name{ForImp-package}
\alias{ForImp-package}
\docType{package}
\title{
Forward Imputation
}
\description{
The package contains a function for the imputation of missing values in matrices of ordinal data, called Forward Imputation,
and other functions for generating ordinal data or imputing missing values.
}
\details{
\tabular{ll}{
Package: \tab ForImp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-24\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alessandro Barbiero<alessandro.barbiero@unimi.it>,
\cr Giancarlo Manzi<giancarlo.manzi@unimi.it>,
\cr Pier Alda Ferrari<pieralda.ferrari@unimi.it>


Maintainer: Alessandro Barbiero<alessandro.barbiero@unimi.it>,
\cr Giancarlo Manzi<giancarlo.manzi@unimi.it>,
\cr Pier Alda Ferrari<pieralda.ferrari@unimi.it>
}
\references{
Ferrari P.A., Annoni P., Barbiero A., Manzi G. (2011) An imputation method for categorical variables with application to nonlinear principal component analysis,
Computational Statistics & Data Analysis, vol. 55, issue 7, pages 2410-2420\cr
\url{http://ideas.repec.org/a/eee/csdana/v55y2011i7p2410-2420.html}\cr
\url{http://www.sciencedirect.com/science/article/pii/S0167947311000521}

Ferrari P.A., Barbiero A., Manzi G.: Handling missing data in presence of ordinal variables: a new imputation procedure. In "New Perspectives in Statistical Modeling and Data Analysis", S. Ingrassia, R. Rocci, M. Vichi, Eds.,  Springer, 2011\cr

Little, R.J.A., Rubin, D.B., 2002. Statistical Analysis with Missing Data, 2nd ed. John Wiley & Sons, Inc.

}
\keyword{ package }
\seealso{
}
\examples{
}
